:bug: Remove custom option when delete a product subscription
