+++ Controller/Adminhtml/RecurrenceProducts/Delete.php
@@ -9,6 +9,7 @@ use Magento\Framework\Registry;
 use Magento\Framework\View\Result\PageFactory;
 use Mundipagg\Core\Recurrence\Services\ProductSubscriptionService;
 use MundiPagg\MundiPagg\Concrete\Magento2CoreSetup;
+use MundiPagg\MundiPagg\Helper\ProductSubscriptionHelper;
 use MundiPagg\MundiPagg\Model\ProductsSubscriptionFactory;
 
 class Delete extends Action
@@ -39,6 +40,8 @@ class Delete extends Action
         $this->messageFactory = $messageFactory;
         Magento2CoreSetup::bootstrap();
 
+        $this->productSubscriptionHelper = new ProductSubscriptionHelper();
+
         parent::__construct($context);
     }
 
@@ -62,6 +65,8 @@ class Delete extends Action
                 return;
             }
         }
+
+        $this->productSubscriptionHelper->deleteRecurrenceCustomOption($productData);
 
         $productSubscriptionService->delete($productId);
